requirements.txt
python-escpos
pywin32
flask
python-dotenv